requirements/base.txt
pandas
sqlalchemy
psycopg2-binary
alembic
